# add benchmark for the cloud output