The benchmark is testing only for httpext.Trail and such that have
different urls and names which is probably the worst case performance
wise. The benchmark is somewhat flaky ... but with more runs it seems
consitant in 5% margins ;)