@MStoykov 2019-11-01
$ benchcmp -mag    9e1c5c11.perf 78858ef4.perf
```
benchmark            old ns/op     new ns/op     delta
BenchmarkCloud-4     19685045      23013543      +16.91%
BenchmarkCloud-4     21060987      19451558      -7.64%
BenchmarkCloud-4     22526058      20867068      -7.36%
BenchmarkCloud-4     20830405      21631604      +3.85%
BenchmarkCloud-4     22701274      22245438      -2.01%

benchmark            old allocs     new allocs     delta
BenchmarkCloud-4     93190          93872          +0.73%
BenchmarkCloud-4     93833          93342          -0.52%
BenchmarkCloud-4     93111          93381          +0.29%
BenchmarkCloud-4     93091          93071          -0.02%
BenchmarkCloud-4     93249          93267          +0.02%

benchmark            old bytes     new bytes     delta
BenchmarkCloud-4     13756303      13806000      +0.36%
BenchmarkCloud-4     13822686      13786573      -0.26%
BenchmarkCloud-4     13790461      13817672      +0.20%
BenchmarkCloud-4     13792657      13819228      +0.19%
BenchmarkCloud-4     13793277      13769954      -0.17%
```

$ benchcmp -mag    78858ef4.perf a948b4df.perf
```
benchmark            old ns/op     new ns/op     delta
BenchmarkCloud-4     23013543      31868265      +38.48%
BenchmarkCloud-4     19451558      26698255      +37.26%
BenchmarkCloud-4     20867068      27512414      +31.85%
BenchmarkCloud-4     21631604      26224220      +21.23%
BenchmarkCloud-4     22245438      25525690      +14.75%

benchmark            old allocs     new allocs     delta
BenchmarkCloud-4     93267          133888         +43.55%
BenchmarkCloud-4     93071          133210         +43.13%
BenchmarkCloud-4     93381          133397         +42.85%
BenchmarkCloud-4     93342          133194         +42.69%
BenchmarkCloud-4     93872          133472         +42.19%

benchmark            old bytes     new bytes     delta
BenchmarkCloud-4     13817672      19473203      +40.93%
BenchmarkCloud-4     13806000      19442325      +40.83%
BenchmarkCloud-4     13769954      19371700      +40.68%
BenchmarkCloud-4     13786573      19379163      +40.57%
BenchmarkCloud-4     13819228      19399568      +40.38%
```
$ benchcmp -mag a948b4df.perf 9554c6e7.perf
```
benchmark            old ns/op     new ns/op     delta
BenchmarkCloud-4     31868265      29050273      -8.84%
BenchmarkCloud-4     26698255      27439261      +2.78%
BenchmarkCloud-4     27512414      28170350      +2.39%
BenchmarkCloud-4     26224220      25671548      -2.11%
BenchmarkCloud-4     25525690      25834813      +1.21%

benchmark            old allocs     new allocs     delta
BenchmarkCloud-4     133210         134313         +0.83%
BenchmarkCloud-4     133194         133753         +0.42%
BenchmarkCloud-4     133888         133707         -0.14%
BenchmarkCloud-4     133397         133275         -0.09%
BenchmarkCloud-4     133472         133428         -0.03%

benchmark            old bytes     new bytes     delta
BenchmarkCloud-4     19371700      19480151      +0.56%
BenchmarkCloud-4     19379163      19419020      +0.21%
BenchmarkCloud-4     19442325      19410276      -0.16%
BenchmarkCloud-4     19399568      19416469      +0.09%
BenchmarkCloud-4     19473203      19466236      -0.04%
```
$ benchcmp -mag  9554c6e7.perf  b6650a55.perf
```
benchmark            old ns/op     new ns/op     delta
BenchmarkCloud-4     25834813      29113323      +12.69%
BenchmarkCloud-4     28170350      26871712      -4.61%
BenchmarkCloud-4     25671548      26679223      +3.93%
BenchmarkCloud-4     27439261      26988018      -1.64%
BenchmarkCloud-4     29050273      28921436      -0.44%

benchmark            old allocs     new allocs     delta
BenchmarkCloud-4     134313         133309         -0.75%
BenchmarkCloud-4     133275         133646         +0.28%
BenchmarkCloud-4     133753         133524         -0.17%
BenchmarkCloud-4     133707         133571         -0.10%
BenchmarkCloud-4     133428         133472         +0.03%

benchmark            old bytes     new bytes     delta
BenchmarkCloud-4     19480151      19378667      -0.52%
BenchmarkCloud-4     19466236      19407682      -0.30%
BenchmarkCloud-4     19419020      19403899      -0.08%
BenchmarkCloud-4     19410276      19413150      +0.01%
BenchmarkCloud-4     19416469      19416639      +0.00%
```