@MStoykov 2019-11-01
@imiric benchmarks are atm not ran by CI although around #1167  may change. 
Performance wise, this is the *worst* case only in it (IMO) we only have HTTPTrails and all of them are in need of copying ... In a real scenario this will likely be 0% of the cases and even where every http request is with `name != url` this will still not be 100% of the metrics ;) In general I expect in somewhat bad cases the performance degradation of the cloud output (not the whole k6) to be around 5-10%