@cuonglm 2019-11-01
Can you retry with `benchstat` instead of `benchcmp`.