@MStoykov 2019-11-01
@cuonglm 
$ benchstat  9e1c5c11.perf 78858ef4.perf a948b4df.perf 9554c6e7.perf
```
name \ time/op    9e1c5c11.perf  78858ef4.perf  a948b4df.perf  9554c6e7.perf
Cloud-4             21.4ms ± 8%    21.4ms ± 9%    27.6ms ±16%    27.2ms ± 7%

name \ alloc/op   9e1c5c11.perf  78858ef4.perf  a948b4df.perf  9554c6e7.perf
Cloud-4             13.8MB ± 0%    13.8MB ± 0%    19.4MB ± 0%    19.4MB ± 0%

name \ allocs/op  9e1c5c11.perf  78858ef4.perf  a948b4df.perf  9554c6e7.perf
Cloud-4              93.3k ± 1%     93.4k ± 1%    133.4k ± 0%    133.7k ± 0%
```