@cuonglm 2019-11-01
> @cuonglm
> $ benchstat [9e1c5c1](https://github.com/loadimpact/k6/commit/9e1c5c11149c699fbbe601d8a724f7a76a869d09).perf [78858ef](https://github.com/loadimpact/k6/commit/78858ef41d2927f43db4f10ca64f791cf7f329ce).perf [a948b4d](https://github.com/loadimpact/k6/commit/a948b4dfc0e1be88a3f48206fe04a6fbf18b6301).perf [9554c6e](https://github.com/loadimpact/k6/commit/9554c6e7e9116351fa330623ac28e3242dacb99c).perf
> 
> ```
> name \ time/op    9e1c5c11.perf  78858ef4.perf  a948b4df.perf  9554c6e7.perf
> Cloud-4             21.4ms ± 8%    21.4ms ± 9%    27.6ms ±16%    27.2ms ± 7%
> 
> name \ alloc/op   9e1c5c11.perf  78858ef4.perf  a948b4df.perf  9554c6e7.perf
> Cloud-4             13.8MB ± 0%    13.8MB ± 0%    19.4MB ± 0%    19.4MB ± 0%
> 
> name \ allocs/op  9e1c5c11.perf  78858ef4.perf  a948b4df.perf  9554c6e7.perf
> Cloud-4              93.3k ± 1%     93.4k ± 1%    133.4k ± 0%    133.7k ± 0%
> ```

Thanks, seems to be a lot of noise when benchmark runs. But the result looks reasonable.